// ETournamentManager.App/src/app/models/branding.model.ts
interface BrandingListingModel {
  platformName: string;
  primaryColor: string;
  secondaryColor: string;
  textColor: string;
  backgroundColor: string;
  accessTeamTable: boolean;
  accessTournamentTable: boolean;
  accessTeamDetails: boolean;
  accessTournamentDetails: boolean;
  contactLink: string;
  contactEmail: string;
}

export default BrandingListingModel;